transfer logout button to account layout
