+++ app/src/main/res/layout/fragment_homepage.xml
@@ -8,24 +8,5 @@
     android:background="@color/teal_200">
 
     <!-- TODO: Update blank fragment layout -->
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:gravity="center">
-            <TextView
-                android:id="@+id/currentUser"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Hello World!"
-                android:textColor="@color/black"
-                />
-
-            <Button
-                android:id="@+id/buttonLogout"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Log out"/>
-        </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
